+++ ViesCraft/1.11.2 - 2206/src/main/java/com/viesis/viescraft/common/items/airshipitems/v1/ItemAirshipV1.java
@@ -34,18 +34,18 @@ public class ItemAirshipV1 extends ItemAirshipCore {
         ItemStack itemstack = playerIn.getHeldItem(handIn);
         
         int balloon = 0;
-        float colorRed = 210;
-		float colorGreen = 180;
-		float colorBlue = 140;
+        int colorRed = 210;
+		int colorGreen = 180;
+		int colorBlue = 140;
     	
         if(playerIn.isSneaking())
 		{
         	if(itemstack.hasTagCompound())
 	        {
         		balloon = itemstack.getTagCompound().getInteger("Balloon"); 
-        		colorRed = itemstack.getTagCompound().getFloat("ColorRed"); 
-				colorGreen = itemstack.getTagCompound().getFloat("ColorGreen"); 
-				colorBlue = itemstack.getTagCompound().getFloat("ColorBlue");
+        		colorRed = itemstack.getTagCompound().getInteger("ColorRed"); 
+				colorGreen = itemstack.getTagCompound().getInteger("ColorGreen"); 
+				colorBlue = itemstack.getTagCompound().getInteger("ColorBlue");
 	        }
         	else
         	{
